rename model package in vod domain to entity package
