+++ domain/vod/src/main/java/entity/Actor.kt
@@ -1,4 +1,4 @@
-package model
+package entity
 
 data class Actor(
     val id: String,
