+++ domain/vod/src/main/java/entity/Language.kt
@@ -1,4 +1,4 @@
-package model
+package entity
 
 enum class Language(val code: String) {
     ENGLISH("en"),
