+++ domain/vod/src/main/java/entity/SearchCategory.kt
@@ -1,4 +1,4 @@
-package model
+package entity
 
 enum class SearchCategory {
     MOVIE,
